Add design matrix standardization and improve LFP preprocessing
Introduces standardize_design_matrix.m to z-score normalize predictors, improving numerical conditioning in GLM fitting. Updates orchestrate_fit_lfp_multichannel.m to high-pass filter LFP data and standardize design matrices per channel, with added diagnostics for data quality. Expands lambda grid and predictor exclusion in defaults.json, removes unused history_basis config, and adds condition number diagnostics to fit_glm_gauss.m. Minor bugfixes and improvements in feature assembly and LFP history block construction.

+++ src/features/assemble_design_matrix.m
@@ -145,7 +145,7 @@ if ~isempty(producedFieldsIncluded)
 end
 
 if ~skipBlock('states')
-    stateBlk = sparse(double([stateConvo, stateSpon]));
+    stateBlk = sparse([double(stateConvo), double(stateSpon)]);
     blockCells{end+1} = stateBlk; %#ok<AGROW>
     stateCols = colStart:(colStart + 1);
     stateMap = struct();
